% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_new.R
\name{box_cox_new}
\alias{box_cox_new}
\title{Box-Cox transform}
\usage{
box_cox_new(formula, mydata, ixx, lambda)
}
\arguments{
\item{formula}{a formula of the form y ~ x + z1 + z2 where y is a binary response variable, x is a continuous predictor variable, and z1, z2, ... are  covariates}

\item{mydata}{dataset used in box cox transform}

\item{ixx}{continuous predictor}

\item{lambda}{lambda used in box cox transform}
}
\value{
data set after transform, contains transformed ixx
}
\description{
This function processes the box cox transform on varibles
}
