% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lboxcox-package.R
\docType{package}
\name{lboxcox-package}
\alias{lboxcox}
\alias{lboxcox-package}
\title{lboxcox: Logistic Box-Cox Regression}
\description{
Implements a logistic Box-Cox model that adds a shape parameter to a
routine logistic regression model, allowing for a flexible relationship
between the log-odds of a binary outcome and a continuous predictor. See
Xing, L. et al. (2021) for the full model description.
}
\section{Fitting a model}{

\code{\link{lbc_maxlik}}, \code{\link{lbc_train_ms}},
\code{\link{lbc_train_bagging}}, and \code{\link{lbc_train_all}} fit the
logistic Box-Cox model with, respectively, a single starting value,
multiple starting values, a bootstrap ensemble around
\code{\link{lbc_maxlik}}, and a bootstrap ensemble around
\code{\link{lbc_train_ms}}. \code{\link{lboxcox_cv.fit}} instead selects
the Box-Cox shape parameter by cross-validated deviance residual.
}

\section{Predicting and interpreting a fitted model}{

\code{\link{lboxcox_maxLik.predict}} and \code{\link{lboxcox_maxLik_el.predict}}
produce predicted probabilities from a fitted model, \code{\link{lboxcox_cv.predict}}
does the same for a \code{\link{lboxcox_cv.fit}} model, \code{\link{median_effect}}
summarizes the fitted exposure-response slope, and \code{\link{devr}} computes
the deviance-residual prediction error used throughout the package.
}

\references{
Xing, L., Zhang, X., Burstyn, I., & Gustafson, P. (2021). On logistic
Box-Cox regression for flexibly estimating the shape and strength of
exposure-disease relationships. Canadian Journal of Statistics, 49(3), 808-825.
}
\author{
\strong{Maintainer}: Li Xing \email{jnjfayaa@gmail.com}

Authors:
\itemize{
  \item Shiyu Xu \email{xushiyu98@outlook.com}
  \item Jing Wang
  \item Kohlton Booth \email{booth.kohl@gmail.com}
  \item Xuekui Zhang
  \item Igor Burstyn
  \item Paul Gustafson
}

}
\keyword{internal}
