% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lboxcox_cv.R
\name{lboxcox_cv.predict}
\alias{lboxcox_cv.predict}
\title{Predict from a \code{lboxcox_cv.fit} model}
\usage{
lboxcox_cv.predict(myCVfit, newdata, formula)
}
\arguments{
\item{myCVfit}{a model fit by \code{\link{lboxcox_cv.fit}}.}

\item{newdata}{data frame of new observations to predict on.}

\item{formula}{the same formula used to fit \code{myCVfit}.}
}
\value{
a numeric vector of predicted probabilities.
}
\description{
Predict from a \code{lboxcox_cv.fit} model
}
