% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikeFun.R
\name{LogLikeFun}
\alias{LogLikeFun}
\title{Log Likelihood of Logistic Box-Cox}
\usage{
LogLikeFun(bb, ixx, iyy, iw, iZZ)
}
\arguments{
\item{bb}{current values for the intercept and slope coefficients}

\item{ixx}{continuous predictor}

\item{iyy}{binary outcome}

\item{iw}{sample weight}

\item{iZZ}{covariates to be incorporated in the model}
}
\value{
the log likelihood estimate for the coefficients in `bb`
}
\description{
This function gives the log likelihood of the Box-Cox model.
Main purpose is to be an input to the maxLik function.
}
