% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreFun.R
\name{ScoreFun}
\alias{ScoreFun}
\title{Log Likelihood Gradient of Logistic Box-Cox}
\usage{
ScoreFun(bb, ixx, iyy, iw, iZZ)
}
\arguments{
\item{bb}{initial values for the intercept and slope coefficients}

\item{ixx}{continuous predictor}

\item{iyy}{binary outcome}

\item{iw}{sample weight}

\item{iZZ}{covariates to be incorporated in the model}
}
\value{
the gradient of the log likelihood estimate for the coefficients in `bb`
}
\description{
This function gives the gradient of the log likelihood of the Box-Cox model.
Main purpose is to be an input to the maxLik function.
}
