% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depress.R
\docType{data}
\name{depress}
\alias{depress}
\title{Depression dataset}
\format{
Sample survey data
\describe{
\item{depression}{binary response variable indicating whether the participant has depression (=1) or not (=0)}
\item{mercury}{a numeric vector giving the log-transformed total blood mercury in micro-grams per litre}
\item{age}{0 of particiapant is female and 1 if they are male}
\item{gender}{age of the participant}
\item{weight}{a numeric vector giving the sampling-weight.}
}
}
\source{
Xing, L., Zhang, X., Burstyn, I., & Gustafson, P. (2021). On logistic Box–Cox regression for flexibly estimating the shape and strength of exposure‐disease relationships. Canadian Journal of Statistics, 49(3), 808-825.
}
\usage{
depress
}
\description{
The depress data frame has 8,893 rows and 5 columns from the National Health and Nutrition Examination Survey (NHANES) 2009–2010.
}
\keyword{datasets}
