% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lboxcox_maxLik_predict.R
\name{lboxcox_maxLik.predict}
\alias{lboxcox_maxLik.predict}
\title{Lboxcox MaxLik Prediction Function}
\usage{
lboxcox_maxLik.predict(myMaxLikfit, newdata, formula)
}
\arguments{
\item{myMaxLikfit}{Fitted model using lboxcox_maxLik model}

\item{newdata}{Given data for prediction}

\item{formula}{a formula of the form y ~ x + z1 + z2 where y is a binary response variable, x is a continuous predictor variable, and z1, z2, ... are  covariates}
}
\value{
p value
}
\description{
Give the predicted p value of given LBC MaxLik model
}
\note{
This is reliant on the following work:
}
