% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lboxcox.R, R/lboxcox_maxLik.R
\name{median_effect}
\alias{median_effect}
\title{Calculates the "slope" of the Logistic Box-Cox model}
\usage{
median_effect(formula, weight_column_name, data, trained_model)

median_effect(formula, weight_column_name, data, trained_model)
}
\arguments{
\item{formula}{the formula used to train the logistic box-cox model}

\item{weight_column_name}{the name of the column in `data` containing the survey weights}

\item{data}{dataframe containing the dataset to train on}

\item{trained_model}{the already trained model. The output of `lbc_train`}
}
\description{
Calculates a number that represents the overall gradient measurement between the predictor and log-odds of the risk

Calculates a number that represents the overall gradient measurement between the predictor and log-odds of the risk
}
